requirements.txt
django==2.2.5
djangorestframework==3.10.3
django-filter==2.2.0
psycopg2==2.8.3